+++ data_provider/__init__.py
@@ -40,6 +40,8 @@ from .yfinance_fetcher import YfinanceFetcher
 from .longbridge_fetcher import LongbridgeFetcher
 from .finnhub_fetcher import FinnhubFetcher
 from .alphavantage_fetcher import AlphaVantageFetcher
+from .finmind_fetcher import FinMindFetcher
+from .tw_market import is_tw_stock_code, canonical_tw_code
 from .us_index_mapping import is_us_index_code, is_us_stock_code, get_us_index_yf_symbol, US_INDEX_MAPPING
 
 __all__ = [
@@ -54,9 +56,12 @@ __all__ = [
     'LongbridgeFetcher',
     'FinnhubFetcher',
     'AlphaVantageFetcher',
+    'FinMindFetcher',
     'is_us_index_code',
     'is_us_stock_code',
     'is_hk_stock_code',
+    'is_tw_stock_code',
+    'canonical_tw_code',
     'get_us_index_yf_symbol',
     'US_INDEX_MAPPING',
 ]